+++ zcash_client_backend/CHANGELOG.md
@@ -182,7 +182,8 @@ and this library adheres to Rust's notion of
     - `WalletTransparentOutput::value`
 
 - `zcash_client_backend::address`:
-  - `RecipientAddress::Shielded` has been renamed to `RecipientAddress::Sapling`
+  - `RecipientAddress` has been renamed to `Address`
+  - `Address::Shielded` has been renamed to `Address::Sapling`
 
 ### Removed
 - `zcash_client_backend::wallet::ReceivedSaplingNote` has been replaced by